callgraph-tool: extract and analyze callgraphs
The callgraph-tool supports to investigate and analyse callgraphs in the
linux kernel.

It includes:

  - hooking into the kernel build process
  - extracting the static callgraph
  - storing it as pickle file for further analysis
  - extending the static callgraph with ftrace data (experimental)
  - some basic queries on the callgraph via command line
  - some user documentation

The current implementation is still very experimental: we are aware of
various known issues and places for improvements in the processing and in
software quality. This needs to be addressed if we consider this tool to
grow with more functionality and querying options.

The scripts are exported from an internal repository with:

    ./callgraph-tool/deploy/deploy-code.py -d $DEPLOY/stage
    ./callgraph-tool/deploy/deploy-data.py -s $DEPLOY/stage -d $DEPLOY/callgraph-deploy
    cp -r $DEPLOY/callgraph-deploy/* $ELISA_REPOSITORY/safety-architecture/tools/callgraph-tool/

executed on safelinux-scripts: 9398de579720ad71ed9d53b8de12c45802fcf0c9

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>