LICENSES: get closer to reuse compliance
As we include the files into subdirectories of the main repository, we
need to move the license files to LICENSES in the top-level directory
of the repository to become reuse-compliant.

$ reuse lint
MISSING COPYRIGHT AND LICENSING INFORMATION

The following files have no copyright and licensing information:
* README.md

SUMMARY

* Bad licenses:
* Deprecated licenses:
* Licenses without file extension:
* Missing licenses:
* Unused licenses:
* Used licenses: Apache-2.0, CC-BY-SA-4.0, GPL-2.0-only
* Read errors: 0
* Files with copyright information: 19 / 20
* Files with license information: 19 / 20

Unfortunately, your project is not compliant with version 3.0 of the REUSE Specification :-(

The content we have added in already reuse compliant, it is just missing
proper copyright information on the README.md that must be added by the
original author.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>