callgraph-tool: graph view: remove node with three dots
It is misleading to draw an edge from a function to node with three dots
(...), since the graph output already includes all the known connected
nodes (for the specified depth-level). Most users would likely interpret the
meaning of three dots as the graph being incomplete for the given depth,
which was clearly not indented.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>