crix-callgraph: use stable kernel v5.9 as an example
Use stable kernel v5.9 as an example in the documentation instead of the
mainline kernel. The main reason is that the mainline kernel no longer
builds with clang-10 version that currently ships with Ubuntu.
We will eventually have to move to newer clang
version, but for now, this is a quick fix to make the commands in the
documentation work as expected.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>