crix-callgraph: fixes after testing on llvm-11 with C++ target
VirtualCallTargets.cc does not currently support llvm-11. This commit
adds compile time checks to make sure VirtualCallTargets.cc is not
used if llvm version is later than 10. In addition, it modifies tests to
make them compatible. Longer-term, VirtualCallTargets.cc should be
modified to support llvm-11, but this is a quick fix for now to allow
using the crix-callgraph against C target programs with llvm-11.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>